Move Input component to ui folder
