+++ src/components/ui/input.tsx
@@ -57,4 +57,4 @@ const Input = React.forwardRef<HTMLInputElement, InputProps>((props, ref) => {
 
 Input.displayName = "Input";
 
-export default Input;
+export { Input };
