+++ src/components/wa/WhatsAppForm.tsx
@@ -2,10 +2,10 @@
 
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useRouter } from "next/navigation";
-import { type SubmitHandler, useForm } from "react-hook-form";
+import { useForm, type SubmitHandler } from "react-hook-form";
 import z from "zod";
-import Input from "../Input";
 import { Button } from "../ui/button";
+import { Input } from "../ui/input";
 
 const formSchema = z.object({
   phoneNumber: z
